Added Signup and login for both mentor and Student
Created login form for mentor and student in the same page.(Both have same details).
Created signup part for both, datas are different.
The data entering part in both are created in separate files.
The scss is created for both in a signle file and imported in signup.scss.
Background images are different both login and signup.
Background color for forms are not good,need to change.
Overall look need to be improved.
Logo given is the old normal logo.